add check uninstall runs from root

+++ linux/uninstall.sh
@@ -1,4 +1,10 @@
 #!/usr/bin/env bash
+
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root" 1>&2
+   exit 1
+fi
+
 echo "Stopping ckb-next app"
 pkill ckb-next >/dev/null 2>&1
 pkill ckb >/dev/null 2>&1
